+++ src/ts/Ads/Managers/UserPrivacyManager.ts
@@ -21,6 +21,7 @@ import { ITemplateData } from 'Core/Views/View';
 import { ConsentPage } from 'Ads/Views/Consent/Consent';
 import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
 import { PrivacySDK } from 'Privacy/PrivacySDK';
+import { PrivacyEvent, PrivacyMetrics } from 'Privacy/PrivacyMetrics';
 
 interface IUserSummary extends ITemplateData {
     deviceModel: string;
@@ -286,17 +287,9 @@ export class UserPrivacyManager {
 
     public setUsersAgeGateChoice(ageGateChoice: AgeGateChoice) {
         if (ageGateChoice === AgeGateChoice.YES) {
-            Diagnostics.trigger('age_gate_pass', {
-                legalFramework: this._privacy.getLegalFramework(),
-                method: this._gamePrivacy.getMethod(),
-                previousChoice: this._ageGateChoice
-            });
+            PrivacyMetrics.trigger(PrivacyEvent.AGE_GATE_PASS);
         } else if (ageGateChoice === AgeGateChoice.NO) {
-            Diagnostics.trigger('age_gate_not_passed', {
-                legalFramework: this._privacy.getLegalFramework(),
-                method: this._gamePrivacy.getMethod(),
-                previousChoice: this._ageGateChoice
-            });
+            PrivacyMetrics.trigger(PrivacyEvent.AGE_GATE_NOT_PASSED);
         }
 
         this._ageGateChoice = ageGateChoice;
@@ -315,6 +308,29 @@ export class UserPrivacyManager {
 
     public getAgeGateChoice(): AgeGateChoice {
         return this._ageGateChoice;
+    }
+
+    public isConsentShowRequired(): boolean {
+        if (this.isAgeGateShowRequired()) {
+            return true;
+        }
+
+        if (!this._gamePrivacy.isEnabled() && this._gamePrivacy.getMethod() !== PrivacyMethod.UNITY_CONSENT) {
+            return false;
+        }
+
+        if (!this._userPrivacy.isRecorded()) {
+            return true;
+        }
+
+        const methodChangedSinceConsent = this._gamePrivacy.getMethod() !== this._userPrivacy.getMethod();
+        const versionUpdatedSinceConsent = this._gamePrivacy.getVersion() > this._userPrivacy.getVersion();
+
+        return methodChangedSinceConsent || versionUpdatedSinceConsent;
+    }
+
+    public getLegalFramework(): LegalFramework {
+        return this._privacy.getLegalFramework();
     }
 
     private pushConsent(consent: boolean): Promise<void> {
@@ -424,5 +440,19 @@ export class UserPrivacyManager {
                 return this._core.Storage.write(StorageType.PRIVATE);
             });
         });
+    }
+
+    private isAgeGateShowRequired(): boolean {
+        if (this._privacy.isAgeGateEnabled()) {
+            if (this._gamePrivacy.getMethod() === PrivacyMethod.LEGITIMATE_INTEREST && this._privacy.isGDPREnabled() && !this._privacy.isOptOutRecorded()) {
+                return true;
+            }
+
+            if (this._gamePrivacy.getMethod() === PrivacyMethod.UNITY_CONSENT && !this._userPrivacy.isRecorded()) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }