+++ src/ts/Core/Utilities/IObserver.ts
@@ -5,3 +5,4 @@ export type IObserver3<P1, P2, P3> = (p1: P1, p2: P2, p3: P3) => void;
 export type IObserver4<P1, P2, P3, P4> = (p1: P1, p2: P2, p3: P3, p4: P4) => void;
 export type IObserver5<P1, P2, P3, P4, P5> = (p1: P1, p2: P2, p3: P3, p4: P4, p5: P5) => void;
 export type IObserver6<P1, P2, P3, P4, P5, P6> = (p1: P1, p2: P2, p3: P3, p4: P4, p5: P5, p6: P6) => void;
+export type IObserver7<P1, P2, P3, P4, P5, P6, P7> = (p1: P1, p2: P2, p3: P3, p4: P4, p5: P5, p6: P6, p7: P7) => void;