+++ src/ts/Performance/AdUnits/PerformanceAdUnitParametersFactory.ts
@@ -14,14 +14,11 @@ import { Campaign } from 'Ads/Models/Campaign';
 import { AbstractPrivacy } from 'Ads/Views/AbstractPrivacy';
 import { AbstractVideoOverlay } from 'Ads/Views/AbstractVideoOverlay';
 import { VideoOverlay } from 'Ads/Views/VideoOverlay';
-import { AnimatedDownloadButtonEndScreen } from 'Performance/Views/AnimatedDownloadButtonEndScreen';
-import {
-    HeartbeatingDownloadButtonTest,
-    BlinkingDownloadButtonTest,
-    BouncingDownloadButtonTest,
-    ShiningDownloadButtonTest
-} from 'Core/Models/ABGroup';
 import { AutomatedExperimentManager } from 'Ads/Managers/AutomatedExperimentManager';
+import { HalloweenThemeFreeTest, FullscreenCTAExperiment } from 'Core/Models/ABGroup';
+import { HalloweenPerformanceEndScreen } from 'Performance/Views/HalloweenPerformanceEndScreen';
+import { VideoOverlayFullscreenCTA } from 'Ads/Views/VideoOverlayFullscreenCTA';
+import { Platform } from 'Core/Constants/Platform';
 
 export class PerformanceAdUnitParametersFactory extends AbstractAdUnitParametersFactory<PerformanceCampaign, IPerformanceAdUnitParameters> {
 
@@ -52,16 +49,10 @@ export class PerformanceAdUnitParametersFactory extends AbstractAdUnitParameters
         const abGroup = baseParams.coreConfig.getAbGroup();
         let endScreen: PerformanceEndScreen;
 
-        if (HeartbeatingDownloadButtonTest.isValid(abGroup)) {
-            endScreen = new AnimatedDownloadButtonEndScreen('heartbeating', endScreenParameters, baseParams.campaign, baseParams.coreConfig.getCountry());
-        } else if (BlinkingDownloadButtonTest.isValid(abGroup)) {
-            endScreen = new AnimatedDownloadButtonEndScreen('blinking', endScreenParameters, baseParams.campaign, baseParams.coreConfig.getCountry());
-        } else if (BouncingDownloadButtonTest.isValid(abGroup)) {
-            endScreen = new AnimatedDownloadButtonEndScreen('bouncing', endScreenParameters, baseParams.campaign, baseParams.coreConfig.getCountry());
-        } else if (ShiningDownloadButtonTest.isValid(abGroup)) {
-            endScreen = new AnimatedDownloadButtonEndScreen('shining', endScreenParameters, baseParams.campaign, baseParams.coreConfig.getCountry());
-        } else {
+        if (HalloweenThemeFreeTest.isValid(abGroup)) {
             endScreen = new PerformanceEndScreen(endScreenParameters, baseParams.campaign, baseParams.coreConfig.getCountry());
+        } else {
+            endScreen = new HalloweenPerformanceEndScreen(endScreenParameters, baseParams.campaign, baseParams.coreConfig.getCountry());
         }
 
         const video = this.getVideo(baseParams.campaign, baseParams.forceOrientation);
@@ -89,7 +80,12 @@ export class PerformanceAdUnitParametersFactory extends AbstractAdUnitParameters
         }
 
         const showGDPRBanner = this.showGDPRBanner(parameters) && showPrivacyDuringVideo;
-        const overlay = new VideoOverlay(parameters, privacy, showGDPRBanner, showPrivacyDuringVideo);
+        let overlay;
+        if (FullscreenCTAExperiment.isValid(parameters.coreConfig.getAbGroup()) && this._platform === Platform.ANDROID) {
+            overlay = new VideoOverlayFullscreenCTA(parameters, privacy, showGDPRBanner, showPrivacyDuringVideo);
+        } else {
+            overlay = new VideoOverlay(parameters, privacy, showGDPRBanner, showPrivacyDuringVideo);
+        }
 
         if (parameters.placement.disableVideoControlsFade()) {
             overlay.setFadeEnabled(false);