+++ test/Unit/MRAID/MRAIDEventHandlersTest.ts
@@ -110,7 +110,7 @@ describe('MRAIDEventHandlersTest', () => {
                 options: {},
                 mraid: mraidView,
                 endScreen: undefined,
-                privacy: new Privacy(platform, extendedMraidCampaign, privacyManager, false, false),
+                privacy: new Privacy(platform, extendedMraidCampaign, privacyManager, false, false, 'en'),
                 privacyManager: privacyManager,
                 programmaticTrackingService: programmaticTrackingService,
                 webPlayerContainer: webPlayerContainer,